IDL: initial minimal parser
This can parse descriptor.proto into the objects defined in
descriptor-pb.dylan, which contains a hand-written version of the code that
will eventually be replaced by the generated code for descriptor.proto. Once
that works we'll be bootstrapped and can start adding all the features that
aren't used in descriptor.proto.