Use generated google-protobuf module definition
This module definition is lightly edited by hand to remove some of the exports
that aren't yet defined, to remove the library definition, and to add three
enum exports that weren't generated due to a bug.

Using the generated module definition required separating all module
definitions out of library.dylan and into their own files, due to the
dependency graph.